requirements.txt
Django==1.8
django-bower==5.1.0
djangorestframework==3.3.2